# internal/api/spec/api-spec.yaml
openapi: 3.0.0
info:
  title: Hyperboard API
  version: 0.1.0
paths:
...
  /api/v1/posts:
    get:
      operationId: getPosts
      summary: Search for posts.
      parameters:
        - $ref: "#/components/parameters/cursor"
        - $ref: "#/components/parameters/limit"
        - $ref: "#/components/parameters/search"
      responses:
        "200":
          $ref: "#/components/responses/PostsResponse"
        "400":
          $ref: "#/components/responses/BadRequestResponse"
        "401":
          $ref: "#/components/responses/UnauthorizedResponse"
        "403":
          $ref: "#/components/responses/ForbiddenResponse"
        "429":
          $ref: "#/components/responses/TooManyRequestsResponse"
        "500":
          $ref: "#/components/responses/InternalServerErrorResponse"
  /api/v1/upload:
    post:
      operationId: uploadPost
      summary: Upload a new post's content.
      parameters:
        - $ref: "#/components/parameters/force"
      requestBody:
        $ref: "#/components/requestBodies/UploadPostRequest"
      responses:
        "201":
          $ref: "#/components/responses/UploadPostResponse"
        "400":
          $ref: "#/components/responses/BadRequestResponse"
        "401":
          $ref: "#/components/responses/UnauthorizedResponse"
        "403":
          $ref: "#/components/responses/ForbiddenResponse"
        "409":
          $ref: "#/components/responses/ConflictResponse"
        "429":
          $ref: "#/components/responses/TooManyRequestsResponse"
        "500":
          $ref: "#/components/responses/InternalServerErrorResponse"
  /api/v1/posts/{id}:
    put:
      operationId: putPost
      summary: Replace a post's metadata. This does not create a new post - use uploadPost to create a new post.
      parameters:
        - $ref: "#/components/parameters/id"
      requestBody:
        $ref: "#/components/requestBodies/PostRequest"
      responses:
        "200":
          $ref: "#/components/responses/PostResponse"
        "400":
          $ref: "#/components/responses/BadRequestResponse"
        "401":
          $ref: "#/components/responses/UnauthorizedResponse"
        "403":
          $ref: "#/components/responses/ForbiddenResponse"
        "404":
          $ref: "#/components/responses/NotFoundResponse"
        "429":
          $ref: "#/components/responses/TooManyRequestsResponse"
        "500":
          $ref: "#/components/responses/InternalServerErrorResponse"
    get:
      operationId: getPost
      summary: Get a post's metadata by ID.
      parameters:
        - $ref: "#/components/parameters/id"
      responses:
        "200":
          $ref: "#/components/responses/PostResponse"
        "400":
          $ref: "#/components/responses/BadRequestResponse"
        "401":
          $ref: "#/components/responses/UnauthorizedResponse"
        "403":
          $ref: "#/components/responses/ForbiddenResponse"
        "404":
          $ref: "#/components/responses/NotFoundResponse"
        "429":
          $ref: "#/components/responses/TooManyRequestsResponse"
        "500":
          $ref: "#/components/responses/InternalServerErrorResponse"
    delete:
      operationId: deletePost
      summary: Delete a post's metadata and content.
      parameters:
        - $ref: "#/components/parameters/id"
      responses:
        "204":
          $ref: "#/components/responses/DeletedResponse"
        "400":
          $ref: "#/components/responses/BadRequestResponse"
        "401":
          $ref: "#/components/responses/UnauthorizedResponse"
        "403":
          $ref: "#/components/responses/ForbiddenResponse"
        "404":
          $ref: "#/components/responses/NotFoundResponse"
        "429":
          $ref: "#/components/responses/TooManyRequestsResponse"
        "500":
          $ref: "#/components/responses/InternalServerErrorResponse"
  /api/v1/posts/{id}/content:
    put:
      operationId: replacePostContent
      summary: Replace a post's content file. The server re-processes the file and updates contentUrl and mimeType.
      parameters:
        - $ref: "#/components/parameters/id"
      requestBody:
        $ref: "#/components/requestBodies/UploadPostRequest"
      responses:
        "200":
          $ref: "#/components/responses/PostResponse"
        "400":
          $ref: "#/components/responses/BadRequestResponse"
        "401":
          $ref: "#/components/responses/UnauthorizedResponse"
        "403":
          $ref: "#/components/responses/ForbiddenResponse"
        "404":
          $ref: "#/components/responses/NotFoundResponse"
        "415":
          $ref: "#/components/responses/BadRequestResponse"
        "422":
          $ref: "#/components/responses/BadRequestResponse"
        "429":
          $ref: "#/components/responses/TooManyRequestsResponse"
        "500":
          $ref: "#/components/responses/InternalServerErrorResponse"
  /api/v1/posts/{id}/thumbnail:
    put:
      operationId: replacePostThumbnail
      summary: Replace a post's thumbnail image. The server processes the image into a WebP thumbnail.
      parameters:
        - $ref: "#/components/parameters/id"
      requestBody:
        $ref: "#/components/requestBodies/UploadPostRequest"
      responses:
        "200":
          $ref: "#/components/responses/PostResponse"
        "400":
          $ref: "#/components/responses/BadRequestResponse"
        "401":
          $ref: "#/components/responses/UnauthorizedResponse"
        "403":
          $ref: "#/components/responses/ForbiddenResponse"
        "404":
          $ref: "#/components/responses/NotFoundResponse"
        "415":
          $ref: "#/components/responses/BadRequestResponse"
        "422":
          $ref: "#/components/responses/BadRequestResponse"
        "429":
          $ref: "#/components/responses/TooManyRequestsResponse"
        "500":
          $ref: "#/components/responses/InternalServerErrorResponse"
...
components:
  schemas:
    Error:
      type: object
      required:
        - message
      properties:
        message:
          type: string
    Cursor:
      type: string
    Limit:
      type: integer
      minimum: 1
      maximum: 1024
      default: 24
    Search:
      type: string
  parameters:
    id:
      name: id
      in: path
      required: true
      schema:
        $ref: "types-spec.yaml#/components/schemas/ID"
...
    cursor:
      name: cursor
      in: query
      required: false
      schema:
        $ref: "#/components/schemas/Cursor"
    limit:
      name: limit
      in: query
      required: false
      schema:
        $ref: "#/components/schemas/Limit"
    search:
      name: search
      in: query
      required: false
      schema:
        $ref: "#/components/schemas/Search"
    force:
      name: force
      in: query
      required: false
      schema:
        type: boolean
        default: false
  requestBodies:
...
    UploadPostRequest:
      required: true
      content:
        application/octet-stream:
          schema:
            type: string
            format: binary
    PostRequest:
      required: true
      content:
        application/json:
          schema:
            $ref: "types-spec.yaml#/components/schemas/Post"
...
  responses:
...
    DeletedResponse:
      description: Resource deleted successfully
    BadRequestResponse:
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    UnauthorizedResponse:
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    ForbiddenResponse:
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    ConflictResponse:
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFoundResponse:
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    TooManyRequestsResponse:
      headers:
        Retry-After:
          schema:
            type: integer
            minimum: 0
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    InternalServerErrorResponse:
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    PostResponse:
      content:
        application/json:
          schema:
            $ref: "types-spec.yaml#/components/schemas/Post"
    UploadPostResponse:
      content:
        application/json:
          schema:
            $ref: "types-spec.yaml#/components/schemas/Post"
    PostsResponse:
      content:
        application/json:
          schema:
            type: object
            properties:
              items:
                type: array
                items:
                  $ref: "types-spec.yaml#/components/schemas/Post"
              cursor:
                $ref: "#/components/schemas/Cursor"
...